reform tracer code
